Multitenant

+++ src/controllers/contacts.ts
@@ -7,15 +7,21 @@ import { success, failure } from '../utils/res'
 import password from '../utils/password'
 import { Op } from 'sequelize'
 import constants from '../constants'
+import * as tribes from '../utils/tribes'
+import * as network from '../network'
 
 export const getContacts = async (req, res) => {
-	const contacts = await models.Contact.findAll({ where: { deleted: false }, raw: true })
-	const invites = await models.Invite.findAll({ raw: true })
-	const chats = await models.Chat.findAll({ where: { deleted: false }, raw: true })
-	const subscriptions = await models.Subscription.findAll({ raw: true })
+	if (!req.owner) return failure(res, 'no owner')
+	const tenant: number = req.owner.id
+
+	const contacts = await models.Contact.findAll({ where: { deleted: false, tenant }, raw: true })
+	const invites = await models.Invite.findAll({ raw: true, where: { tenant } })
+	const chats = await models.Chat.findAll({ where: { deleted: false, tenant }, raw: true })
+	const subscriptions = await models.Subscription.findAll({ raw: true, where: { tenant } })
 	const pendingMembers = await models.ChatMember.findAll({
 		where: {
-			status: constants.chat_statuses.pending
+			status: constants.chat_statuses.pending,
+			tenant
 		}
 	})
 
@@ -49,7 +55,14 @@ export const getContacts = async (req, res) => {
 export const generateToken = async (req, res) => {
 	console.log('=> generateToken called', { body: req.body, params: req.params, query: req.query })
 
-	const owner = await models.Contact.findOne({ where: { isOwner: true } })
+	const pubkey = req.body['pubkey']
+	if(!pubkey) {
+		return failure(res, 'no pubkey')
+	}
+	const owner = await models.Contact.findOne({ where: { isOwner: true, publicKey: pubkey } })
+	if(!owner) {
+		return failure(res, 'no owner')
+	}
 
 	const pwd = password
 	if (process.env.USE_PASSWORD === 'true') {
@@ -61,33 +74,33 @@ export const generateToken = async (req, res) => {
 		}
 	}
 
-	if (owner) {
-		const token = req.body['token']
-		if(!token) {
+	const token = req.body['token']
+	if (!token) {
+		return failure(res, {})
+	}
+	const hash = crypto.createHash('sha256').update(token).digest('base64');
+
+	if (owner.authToken) {
+		if (owner.authToken !== hash) {
 			return failure(res, {})
 		}
-		const hash = crypto.createHash('sha256').update(token).digest('base64');
-
-		if (owner.authToken) {
-			if (owner.authToken!==hash) {
-				return failure(res, {})
-			}
-		} else {
-			owner.update({ authToken: hash })
-		}
-
-		success(res, {})
 	} else {
-		failure(res, {})
+		// done!
+		tribes.subscribe(`${pubkey}/#`, network.receiveMqttMessage) // add MQTT subsription
+		owner.update({ authToken: hash })
 	}
+
+	success(res, {id: (owner && owner.id) || 0})
 }
 
 export const updateContact = async (req, res) => {
+	if (!req.owner) return failure(res, 'no owner')
+	const tenant: number = req.owner.id
 	console.log('=> updateContact called', { body: req.body, params: req.params, query: req.query })
 
 	let attrs = extractAttrs(req.body)
 
-	const contact = await models.Contact.findOne({ where: { id: req.params.id } })
+	const contact = await models.Contact.findOne({ where: { id: req.params.id, tenant } })
 	if (!contact) {
 		return failure(res, 'no contact found')
 	}
@@ -109,8 +122,8 @@ export const updateContact = async (req, res) => {
 	}
 
 	// send updated owner info to others!
-	const contactIds = await models.Contact.findAll({ where: { deleted: false } })
-		.filter(c => c.id !== 1 && c.publicKey).map(c => c.id)
+	const contactIds = await models.Contact.findAll({ where: { deleted: false, tenant } })
+		.filter(c => c.id !== tenant && c.publicKey).map(c => c.id)
 	if (contactIds.length == 0) return
 
 	console.log("=> send contact_key to", contactIds)
@@ -123,10 +136,12 @@ export const updateContact = async (req, res) => {
 }
 
 export const exchangeKeys = async (req, res) => {
+	if (!req.owner) return failure(res, 'no owner')
+	const tenant: number = req.owner.id
 	console.log('=> exchangeKeys called', { body: req.body, params: req.params, query: req.query })
 
-	const contact = await models.Contact.findOne({ where: { id: req.params.id } })
-	const owner = await models.Contact.findOne({ where: { isOwner: true } })
+	const contact = await models.Contact.findOne({ where: { id: req.params.id, tenant } })
+	const owner = req.owner
 
 	success(res, jsonUtils.contactToJson(contact))
 
@@ -138,13 +153,15 @@ export const exchangeKeys = async (req, res) => {
 }
 
 export const createContact = async (req, res) => {
+	if (!req.owner) return failure(res, 'no owner')
+	const tenant: number = req.owner.id
 	console.log('=> createContact called', { body: req.body, params: req.params, query: req.query })
 
 	let attrs = extractAttrs(req.body)
 
-	const owner = await models.Contact.findOne({ where: { isOwner: true } })
+	const owner = req.owner
 
-	const existing = attrs['public_key'] && await models.Contact.findOne({ where: { publicKey: attrs['public_key'] } })
+	const existing = attrs['public_key'] && await models.Contact.findOne({ where: { publicKey: attrs['public_key'], tenant } })
 	if (existing) {
 		const updateObj: { [k: string]: any } = { fromGroup: false }
 		if (attrs['alias']) updateObj.alias = attrs['alias']
@@ -153,6 +170,7 @@ export const createContact = async (req, res) => {
 	}
 
 	if (attrs['public_key'].length > 66) attrs['public_key'] = attrs['public_key'].substring(0, 66)
+	attrs.tenant = tenant
 	const createdContact = await models.Contact.create(attrs)
 	const contact = await createdContact.update(jsonUtils.jsonToContact(attrs))
 
@@ -166,21 +184,23 @@ export const createContact = async (req, res) => {
 }
 
 export const deleteContact = async (req, res) => {
+	if (!req.owner) return failure(res, 'no owner')
+	const tenant: number = req.owner.id
 	const id = parseInt(req.params.id || '0')
-	if (!id || id === 1) {
+	if (!id || id === tenant) {
 		failure(res, 'Cannot delete self')
 		return
 	}
 
-	const contact = await models.Contact.findOne({ where: { id } })
+	const contact = await models.Contact.findOne({ where: { id, tenant } })
 	if (!contact) return
 
-	const owner = await models.Contact.findOne({ where: { isOwner: true } })
-	const tribesImAdminOf = await models.Chat.findAll({ where: { ownerPubkey: owner.publicKey } })
+	const owner = req.owner
+	const tribesImAdminOf = await models.Chat.findAll({ where: { ownerPubkey: owner.publicKey, tenant } })
 	const tribesIdArray = tribesImAdminOf && tribesImAdminOf.length && tribesImAdminOf.map(t => t.id)
 	let okToDelete = true
 	if (tribesIdArray && tribesIdArray.length) {
-		const thisContactMembers = await models.ChatMember.findAll({ where: { contactId: id, chatId: { [Op.in]: tribesIdArray } } })
+		const thisContactMembers = await models.ChatMember.findAll({ where: { contactId: id, chatId: { [Op.in]: tribesIdArray }, tenant } })
 		if (thisContactMembers && thisContactMembers.length) {
 			// IS A MEMBER! dont delete, instead just set from_group=true
 			okToDelete = false
@@ -199,7 +219,7 @@ export const deleteContact = async (req, res) => {
 	}
 
 	// find and destroy chat & messages
-	const chats = await models.Chat.findAll({ where: { deleted: false } })
+	const chats = await models.Chat.findAll({ where: { deleted: false, tenant } })
 	chats.map(async chat => {
 		if (chat.type === constants.chat_types.conversation) {
 			const contactIds = JSON.parse(chat.contactIds)
@@ -210,45 +230,51 @@ export const deleteContact = async (req, res) => {
 					contactIds: '[]',
 					name: ''
 				})
-				await models.Message.destroy({ where: { chatId: chat.id } })
+				await models.Message.destroy({ where: { chatId: chat.id, tenant } })
 			}
 		}
 	})
-	await models.Invite.destroy({ where: { contactId: id } })
-	await models.Subscription.destroy({ where: { contactId: id } })
+	await models.Invite.destroy({ where: { contactId: id, tenant } })
+	await models.Subscription.destroy({ where: { contactId: id, tenant } })
 
 	success(res, {})
 }
 
 export const receiveContactKey = async (payload) => {
-	console.log('=> received contact key', JSON.stringify(payload))
 
 	const dat = payload.content || payload
 	const sender_pub_key = dat.sender.pub_key
+	const sender_route_hint = dat.sender.route_hint
 	const sender_contact_key = dat.sender.contact_key
 	const sender_alias = dat.sender.alias || 'Unknown'
 	const sender_photo_url = dat.sender.photo_url
+	const owner = payload.owner
+	const tenant: number = owner.id
+
+	console.log('=> received contact key from',  sender_pub_key, tenant)
 
 	if (!sender_pub_key) {
 		return console.log("no pubkey!")
 	}
 
-	const owner = await models.Contact.findOne({ where: { isOwner: true } })
-	const sender = await models.Contact.findOne({ where: { publicKey: sender_pub_key, status: constants.contact_statuses.confirmed } })
+	const sender = await models.Contact.findOne({ where: { publicKey: sender_pub_key, status: constants.contact_statuses.confirmed, tenant } })
 	let msgIncludedContactKey = false // ???????
-	if(sender_contact_key) {
+	if (sender_contact_key) {
 		msgIncludedContactKey = true
 	}
-	if (sender_contact_key && sender) {
-		const objToUpdate: { [k: string]: any } = { contactKey: sender_contact_key }
+	if (sender) {
+		const objToUpdate: { [k: string]: any } = { }
+		if (sender_contact_key) objToUpdate.contactKey = sender_contact_key
 		if (sender_alias) objToUpdate.alias = sender_alias
 		if (sender_photo_url) objToUpdate.photoUrl = sender_photo_url
-		await sender.update(objToUpdate)
-
+		if(Object.keys(objToUpdate).length) {
+			await sender.update(objToUpdate)
+		}
+		
 		socket.sendJson({
 			type: 'contact',
 			response: jsonUtils.contactToJson(sender)
-		})
+		}, tenant)
 	} else {
 		console.log("DID NOT FIND SENDER")
 	}
@@ -256,6 +282,7 @@ export const receiveContactKey = async (payload) => {
 	if (msgIncludedContactKey) {
 		helpers.sendContactKeys({
 			contactPubKey: sender_pub_key,
+			routeHint: sender_route_hint,
 			contactIds: sender ? [sender.id] : [],
 			sender: owner,
 			type: constants.message_types.contact_key_confirmation,
@@ -271,12 +298,14 @@ export const receiveConfirmContactKey = async (payload) => {
 	const sender_contact_key = dat.sender.contact_key
 	const sender_alias = dat.sender.alias || 'Unknown'
 	const sender_photo_url = dat.sender.photo_url
+	const owner = dat.owner
+	const tenant: number = owner.id
 
 	if (!sender_pub_key) {
 		return console.log("no pubkey!")
 	}
 
-	const sender = await models.Contact.findOne({ where: { publicKey: sender_pub_key, status: constants.contact_statuses.confirmed } })
+	const sender = await models.Contact.findOne({ where: { publicKey: sender_pub_key, status: constants.contact_statuses.confirmed, tenant } })
 	if (sender_contact_key && sender) {
 		const objToUpdate: { [k: string]: any } = { contactKey: sender_contact_key }
 		if (sender_alias) objToUpdate.alias = sender_alias
@@ -286,12 +315,12 @@ export const receiveConfirmContactKey = async (payload) => {
 		socket.sendJson({
 			type: 'contact',
 			response: jsonUtils.contactToJson(sender)
-		})
+		}, tenant)
 	}
 }
 
-const extractAttrs = body => {
-	let fields_to_update = ["public_key", "node_alias", "alias", "photo_url", "device_id", "status", "contact_key", "from_group", "private_photo", "notification_sound", "tip_amount"]
+function extractAttrs(body): { [k: string]: any } {
+	let fields_to_update = ["public_key", "node_alias", "alias", "photo_url", "device_id", "status", "contact_key", "from_group", "private_photo", "notification_sound", "tip_amount", "route_hint"]
 	let attrs = {}
 	Object.keys(body).forEach(key => {
 		if (fields_to_update.includes(key)) {
